==> Step2_EDA_Data_Transformation_n_feature_engineering_shoppers.ipynb <==
## Objective : In this Notebook we will perform data transformation as per requirement

import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
import sidetable

import warnings
warnings.filterwarnings('ignore')
%matplotlib inline

from sklearn.pipeline import Pipeline
from sklearn.model_selection import StratifiedShuffleSplit
from sklearn.preprocessing import MinMaxScaler , StandardScaler 
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline

%load_ext autoreload
%autoreload 2

### Import Function File

import Functions_File as my_ff

#  Transform  Data

train_data=pd.read_csv(r'C:\Users\psahu\Documents\My_projects\online_shoppers\online_shoppers_intention_modified.csv' )

train_data.shape

train_data.columns

Note : We will  treat below columns as categorical features:
'SpecialDay', 'OperatingSystems', 'Browser', 'Region', 'TrafficType'

train_data[['SpecialDay', 'OperatingSystems', 'Browser', 'Region', 'TrafficType','Revenue']] = train_data[['SpecialDay', 'OperatingSystems', 'Browser', 'Region', 'TrafficType','Revenue']].astype('str')

#Drop Co-related Features
train_data.drop(['Administrative',  'Informational',  'ProductRelated',  'BounceRates'] ,axis =1 ,inplace = True)

train_data.columns

### Save Target Feature

Target_var = train_data['Revenue']
train_data.drop('Revenue' ,axis=1 ,inplace = True)

### Feature Engineering 

#Check columns types (change below features into  categorical)

train_data[['SpecialDay', 'OperatingSystems', 'Browser', 'Region', 'TrafficType']] = train_data[['SpecialDay', 'OperatingSystems', 'Browser', 'Region', 'TrafficType']].astype('str')

numerical_feats ,categorical_feats = my_ff.column_type(train_data)

print('-------------------------------------------')
print('Numeric Features:'  , list(numerical_feats))
print('-------------------------------------------')
print('Categorical Features:'  , list(categorical_feats))

### Dummy Encoding of Categorical Variable

Encoded_vars_train = pd.get_dummies(train_data[categorical_feats] , drop_first =True)

train_data = pd.concat([train_data ,Encoded_vars_train ],axis =1)

train_data.drop(categorical_feats , axis =1 , inplace = True)
train_data

### Split and save Data

y = Target_var
X = train_data

#Check : stratify = y for good class distribution
import sklearn.model_selection as model_selection
X_train, X_test, y_train, y_test = model_selection.train_test_split(
    X, y, test_size=0.2, random_state=1 , stratify = y)
X_train, X_val, y_train, y_val = model_selection.train_test_split(
    X_train, y_train, test_size=0.25, random_state=1 )
 

print(y_train.value_counts(normalize=True)) 
print(y_val.value_counts(normalize=True)) 
print(y_test.value_counts(normalize=True))

X_train.shape , X_test.shape , X_val.shape , y_train.shape , y_test.shape , y_val.shape 

#This data is ready for Tree - based algorithms

X_train.to_csv(r'C:\Users\psahu\Documents\My_projects\online_shoppers\data\Transformed_X_train_shoppersData.csv' , index =False)
X_test.to_csv(r'C:\Users\psahu\Documents\My_projects\online_shoppers\data\Transformed_X_test_shoppersData.csv' , index =False)
X_val.to_csv(r'C:\Users\psahu\Documents\My_projects\online_shoppers\data\Transformed_X_val_shoppersData.csv' , index =False)
y_train.to_csv(r'C:\Users\psahu\Documents\My_projects\online_shoppers\data\Transformed_y_train.csv' , index =False)
y_test.to_csv(r'C:\Users\psahu\Documents\My_projects\online_shoppers\data\Transformed_y_test.csv' , index =False)
y_val.to_csv(r'C:\Users\psahu\Documents\My_projects\online_shoppers\data\Transformed_y_val.csv' , index =False)
