==> Step3_Trying_Multiple Models_shoppers.ipynb <==
### Objective : In this Notebook we will choose the best performing model out of various models and fine tune it

#import libraries here; add more as necessary.
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import time
import math
import warnings
warnings.filterwarnings("ignore")

#Models
from sklearn.linear_model import LogisticRegression 
from sklearn.svm import SVC
from sklearn.neighbors import KNeighborsClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import AdaBoostClassifier
from sklearn.ensemble import GradientBoostingClassifier
from xgboost import XGBClassifier
from sklearn.feature_selection import RFE , RFECV

from sklearn.model_selection import GridSearchCV
from imblearn.pipeline import Pipeline
from imblearn.over_sampling import SMOTE
from sklearn.model_selection import RepeatedStratifiedKFold
from sklearn.model_selection import cross_val_score
from pandas import ExcelWriter
from openpyxl import load_workbook

## Load Data

X_train = pd.read_csv(r'C:\Users\psahu\Documents\My_projects\online_shoppers\data\Transformed_X_train_shoppersData.csv')
y_train = pd.read_csv(r'C:\Users\psahu\Documents\My_projects\online_shoppers\data\Transformed_y_train.csv')

X_train.shape , y_train.shape

## Declare Important Functions and variables

lr  = LogisticRegression()
svc = SVC(random_state = 50)
knn = KNeighborsClassifier()
rfc = RandomForestClassifier(random_state = 50)
abc = AdaBoostClassifier(random_state=50) # AdaBoostClassifier
gbc = GradientBoostingClassifier(random_state=50) # GradientBoostingClassifier
xgb = XGBClassifier(random_state = 50)


clf_list = [lr, knn,rfc,abc , gbc, xgb]

# Multi Model Performance

print(X_train.shape)
y_train.shape

y_train.value_counts()

def fit_model( clf , param_grid, X=X_train, y=y_train):

    
    scorer = { 
        'accuracy'   : 'accuracy',
         'roc_auc'   : 'roc_auc',
            'f1'     : 'f1',
         'precision' : 'precision',
            'recall' : 'recall'    
                }
    
    myratio = {1: round( 1 *  y_train['Revenue'].value_counts()[0],0).astype(int) } 
    clf_model = Pipeline(
        [
      ('sampling', SMOTE(sampling_strategy= myratio )),
        ('clf', clf)
    ])

    # cv uses StratifiedKFold
    # scoring r2_score available as parameter
    start = time.time()
    grid = GridSearchCV(estimator=clf_model, param_grid= {}, scoring= scorer ,  cv=5 , refit ='recall')
    print("Training {} :".format(clf.__class__.__name__))
    grid.fit(X, y)
    end = time.time()
    time_taken = round((end-start)/60,2)
    #Recall_score = round(grid.best_score_,4)
    
    accuracy = grid.cv_results_['mean_test_accuracy']
    roc_auc = grid.cv_results_['mean_test_roc_auc']
    f1 = grid.cv_results_['mean_test_f1']
    precision = grid.cv_results_['mean_test_precision']
    recall = grid.cv_results_['mean_test_recall']


    print(clf.__class__.__name__)
    print("Time taken : {} mins".format(time_taken))
    print("Accuracy : {}".format(accuracy))
    print("*"*40)
    
    return accuracy, roc_auc, f1, precision, recall, time_taken

# Find best classification algorithm

clf_names = []
clf_accuracy = []
clf_roc_auc = []
clf_f1 = []
clf_precision = []
clf_recall = []
#clf_scores = []
#clf_best_ests = []
clf_time_taken = []
clf_dict = {}

for clf in clf_list:
    accuracy, roc_auc, f1, precision, recall, time_taken = fit_model(clf, {})
    clf_names.append(clf.__class__.__name__)
    clf_accuracy.append(accuracy)
    clf_roc_auc.append(roc_auc)
    clf_f1.append(f1)
    clf_precision.append(precision)
    clf_recall.append(recall)
#   clf_scores.append(best_score)
#   clf_best_ests.append(best_est)
    clf_time_taken.append(time_taken)

#Round 
clf_accuracy = list(np.round(clf_accuracy,5))
clf_roc_auc = list(np.round(clf_roc_auc,5))
clf_f1 = list(np.round(clf_f1,5))
clf_precision = list(np.round(clf_precision,5))
clf_recall = list(np.round(clf_recall,5))

# Create clf_df dataframe from clf_dict
clf_dict['Accuracy'] = clf_accuracy
clf_dict['Roc_Auc'] = clf_roc_auc
clf_dict['F1'] =   clf_f1
clf_dict['Precision'] = clf_precision
clf_dict['Recall'] = clf_recall
clf_dict['Time_Taken (mins)'] = clf_time_taken
#clf_dict['roc_auc'] = clf_scores
#clf_dict['best_est'] = clf_best_ests
clf_df = pd.DataFrame(clf_dict, index=clf_names)
clf_df

# Get the best Hyper-Parameters using Grid Search

#!pip install pactools

from sklearn.metrics import  make_scorer
from imblearn.pipeline import Pipeline
from pactools.grid_search import GridSearchCVProgressBar

clf = XGBClassifier(random_state = 50)

def fit_classifier(clf, param_grid,  X=X_train, y=y_train): 
    
    scorer = {'recall' : 'recall' , 
              'accuracy' : 'accuracy'}
    
    myratio = {1: round( 1 *  y_train['Revenue'].value_counts()[0],0).astype(int) }
    
    clf_model = Pipeline(  [
        ('sampling', SMOTE(sampling_strategy= myratio )),
        ('clf', clf)
    ])

    start = time.time()
    grid = GridSearchCV(clf_model, param_grid=param_grid, cv = 5 ,scoring= scorer, verbose = 10 , refit ='recall', n_jobs=-1 )
    print("Training {} :".format(clf.__class__.__name__))
    grid.fit(X, y)
    end = time.time()
    time_taken = round((end-start)/3600,2)

    print(clf.__class__.__name__)
    print("Time taken : {} hours".format(time_taken))
    print("Best score : {}".format(round(grid.best_score_,4)))
    print("*"*40)
    
    return grid.best_score_, grid.best_estimator_, time_taken

param_grid = {
 "clf__learning_rate"    : [0.05, 0.1, 0.3  ] ,
 'clf__n_estimators'     : [100 , 200 , 300],
 "clf__max_depth"        : [4, 5 ,6, 7],
 "clf__min_child_weight" : [1, 2 , 3 ],
 "clf__gamma"            : [0.0 ,0.1,0.2],
 "clf__colsample_bytree" : [0.5 , 0.7, 1.0]
}

xgb_best_score, xgb_best_est, time_taken = fit_classifier(XGBClassifier(random_state = 50), param_grid , X_train, y_train)
xgb_best_est

print("Best: %f using %s in %f hours" % (xgb_best_score, xgb_best_est , time_taken))
