==> Step4_Final_Model_with_decile_Online_shoppers.ipynb <==
## Objective : We will measure performance of final model .

#!pip install  scikit-plot

#import libraries here; add more as necessary.
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import time
import sidetable
import warnings
import joblib
warnings.filterwarnings("ignore")

#Models
from xgboost import XGBClassifier


from imblearn.over_sampling import SMOTE
from sklearn.model_selection import GridSearchCV
from sklearn.preprocessing import MinMaxScaler
from sklearn.metrics import accuracy_score, confusion_matrix
from pandas import ExcelWriter
from openpyxl import load_workbook

%load_ext autoreload
%autoreload 2

## Load Data

X_train = pd.read_csv(r'C:\Users\psahu\Documents\My_projects\online_shoppers\data\Transformed_X_train_shoppersData.csv')
y_train = pd.read_csv(r'C:\Users\psahu\Documents\My_projects\online_shoppers\data\Transformed_y_train.csv')

X_test = pd.read_csv(r'C:\Users\psahu\Documents\My_projects\online_shoppers\data\Transformed_X_test_shoppersData.csv')
y_test = pd.read_csv(r'C:\Users\psahu\Documents\My_projects\online_shoppers\data\Transformed_y_test.csv')

X_val = pd.read_csv(r'C:\Users\psahu\Documents\My_projects\online_shoppers\data\Transformed_X_val_shoppersData.csv')
y_val = pd.read_csv(r'C:\Users\psahu\Documents\My_projects\online_shoppers\data\Transformed_y_val.csv')

import Functions_File as my_ff

# Final Model

print(X_train.shape ,y_train.shape)

print(X_test.shape ,y_test.shape)

print(X_val.shape ,y_val.shape)

## Before Sampling

y_train.value_counts()

#USing SMOTE for oversampling of Minority class
smt = SMOTE(sampling_strategy = 1)
X_train, y_train = smt.fit_sample(X_train, y_train)

## After Sampling

y_train.value_counts()

## XGBoost

Outcome of Hyper-parameter Tuning
XGBClassifier(colsample_bytree=0.7, gamma=0.0,
                               learning_rate=0.05, max_depth=4,
                               min_child_weight=3, random_state=50)

#model =  XGBClassifier()

model =  XGBClassifier(colsample_bytree=0.7, gamma=0.0,
                               learning_rate=0.05, max_depth=4,
                               min_child_weight=3, random_state=50)

model.fit(X_train,y_train)

## Prediction on test set.  

y_pred_test = model.predict(X_test)
print("Accuracy:" , accuracy_score(y_test,y_pred_test)*100)

y_pred_test_prob=model.predict_proba(X_test)

confusion_matrix(y_test, y_pred_test)

from sklearn.metrics import recall_score
recall_score(y_test, y_pred_test,  average='binary')

#Compute precision, recall, F-measure and support

from sklearn.metrics import classification_report
print(classification_report(y_test, y_pred_test))

##  Prediction on validation set 

y_pred_val = model.predict(X_val)
print("Accuracy:" , accuracy_score(y_val,y_pred_val)*100)

recall_score_val=recall_score(y_val,y_pred_val,  average='binary')
recall_score_val

y_pred_val_proba =model.predict_proba(X_val)

#
X_val_probs = model.predict_proba(X_val)[:,1]
#%% Confusion metrics - validation set
confusion_matrix(y_val, y_pred_val)

#%% Compute precision, recall, F-measure and support

from sklearn.metrics import classification_report
print(classification_report(y_val, y_pred_val))

## Feature Importance

xgb_feat_imp_cover = model.get_booster().get_score(importance_type='cover')
xgb_feat_imp_gain = model.get_booster().get_score(importance_type='gain')
xgb_feat_imp_weight = model.get_booster().get_score(importance_type='weight')

full_feat_imp = pd.DataFrame( [ xgb_feat_imp_gain , xgb_feat_imp_cover, xgb_feat_imp_weight]).T
full_feat_imp.columns = [ 'gain' ,'cover' ,  'weight']

#Sort as per 'gain'
full_feat_imp.sort_values("gain", axis = 0, ascending = False , inplace = True)

full_feat_imp=full_feat_imp.reset_index().rename(columns = {'index' : 'features'})
full_feat_imp

## Decile

#Decile is implemented on Validation dataset here ,implement it  on scoring dataset in real world scenario.
#Just need to add Unique ID in Final Output.

def decile_func(X):
    if      (X >= 0 and X <.10) : return '10'               
    elif          (X >=.10 and X <.20): return '9'                   
    elif         (X>=.20 and X <.30): return '8'                   
    elif         (X>=.30 and X <.40): return '7'                    
    elif         (X>=.40 and X <.50): return '6'                   
    elif         (X>=.50 and X <.60): return '5'                    
    elif         (X>=.60 and X <.70): return '4'                   
    elif         (X>=.70 and X <.80): return '3'                   
    elif         (X>=.80 and X <.90): return '2'                   
    elif         (X>=.90 and X <=1): return '1'  
    else : return 'NA'


y_pred_val_proba[:,0]

df_proba = pd.DataFrame(y_pred_val_proba[:,0] , columns = ['Probability'])

df_proba.tail()


#Apply Decile Function
df_proba['Decile'] =  df_proba['Probability'].apply(decile_func)

df_proba

df_proba['Decile'].value_counts()

df_proba['Decile'].value_counts(normalize=True).mul(100).round(2).astype('str') + '%'
